requirements.txt
affine==2.3.0
astroid==2.4.2
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
contextily==1.0.1
cycler==0.10.0
Fiona==1.8.18
geographiclib==1.50
geopandas==0.8.1
geopy==2.0.0
idna==2.10
isort==5.6.4
joblib==1.0.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.3
mccabe==0.6.1
mercantile==1.1.6
munch==2.5.0
numpy==1.19.4
pandas==1.1.5
Pillow==8.0.1
psutil==5.7.3
pylint==2.6.0
pyparsing==2.4.7
pyproj==3.0.0.post1
PyQt5==5.15.2
PyQt5-sip==12.8.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.4
rasterio==1.1.8
requests==2.25.1
Shapely==1.7.1
six==1.15.0
snuggs==1.4.7
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
